Feature/wmsdk 455 parsing inapp quota (#539)
* Adds InAppSettings model and parsing

Adds a new `InAppSettingsModel` to represent in-app message settings.

Extends the `SettingsModel` to include an optional `inapp` property of type `InAppSettings`.

Adds JSON stubs and tests to validate the parsing of the new in-app settings. This includes tests for valid configurations, missing values, and type errors to ensure robust parsing and handling of different config scenarios.

The changes enable the configuration of in-app message behavior through the config.

* Adds in-app settings logging

* WMSDK-455 PR Fix

---------

Co-authored-by: Vailence <utekeshev@mindbox.cloud>